Crear, Editar y borrar citas

+++ src/components/CitaTableRow.js
@@ -11,25 +11,31 @@ export default class CitaTableRow extends Component {
     }
 
     borrarCita() {
-        axios.delete('http://localhost:4000/citas/borrar-cita/' + this.props.obj._id)
-            .then((res) => {
-                console.log('Cita borrada con exito!')
-                window.location.reload();
-            }).catch((error) => {
-                console.log(error)
-            })
+        if(window.confirm('Esta seguro de borrar la cita?')){
+            axios.delete('http://localhost:4000/citas/borrar-cita/' + this.props.obj._id)
+                .then((res) => {
+                    console.log('Cita borrada con exito!')
+                    window.location.reload();
+                }).catch((error) => {
+                    console.log(error)
+                })
+        }
         
     }
 
     render() {
-
+        //const esPlanta = this.props.obj.tipo === 'Planta'
+        const esMecanico = this.props.obj.tipo === 'Mecanico'
+        const usuarioAdmin = JSON.parse(localStorage.getItem('token')).usuario.tipo
 
         return (
             <tr>
                 <td>{this.props.obj.servicio.nombre}</td>
-                <td>{this.props.obj.servicio.mecanico.label}</td>
+                <td>{this.props.obj.mecanico.nombre}</td>
                 <td>{this.props.obj.servicio.duracionhoras}</td>
                 <td>{this.props.obj.estado}</td>
+                <td>{this.props.obj.cliente}</td>
+                <td>{this.props.obj.telefono}</td>                
                 <td>{this.props.obj.fecha.split('T')[0]}</td>
                 <td>{this.props.obj.hora+":00"}</td>
                 <td>{this.props.obj.placavehiculo}</td>
@@ -43,7 +49,9 @@ export default class CitaTableRow extends Component {
                     <Link className="btn btn-warning" to={"/comentarios/" + this.props.obj._id}>
                         Comentarios
                     </Link>
-                    <Button onClick={this.borrarCita} className="btn btn-danger" variant="danger">Borrar</Button>
+                    {usuarioAdmin ? (
+                        <Button onClick={this.borrarCita} className="btn btn-danger" variant="danger">Borrar</Button>
+                    ):(<div></div>)}
                 </td>
             </tr>
         );
